+++ src/main/java/module-info.java
@@ -18,15 +18,15 @@
  * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-module gradle.migration.data.main
+module io.github.astrapisixtynine.gradle.migration.data
 {
-	requires lombok;
+	requires static lombok;
 	requires org.apache.commons.lang3;
 	requires org.apache.commons.text;
-	requires silly.collection;
-	requires file.worker;
-	requires silly.io.main;
-	requires silly.strings;
+	requires io.github.astrapisixtynine.file.worker;
+	requires io.github.astrapisixtynine.silly.collection;
+	requires io.github.astrapisixtynine.silly.io;
+	requires io.github.astrapisixtynine.silly.strings;
 	requires java.logging;
 
 	exports io.github.astrapi69.gradle.migration.extension;